+++ src/openms/include/OpenMS/ANALYSIS/ID/AScore.h
@@ -13,7 +13,7 @@
 #include <OpenMS/DATASTRUCTURES/String.h>
 #include <OpenMS/KERNEL/StandardTypes.h>
 #include <OpenMS/KERNEL/MSSpectrum.h>
-#include <OpenMS/ANALYSIS/RNPXL/PScore.h>
+#include <OpenMS/ANALYSIS/ID/PScore.h>
 
 #include <limits>
 #include <vector>
