+++ src/openms/include/OpenMS/ANALYSIS/RNPXL/sources.cmake
@@ -3,9 +3,6 @@ set(directory include/OpenMS/ANALYSIS/RNPXL)
 
 ### list all header files of the directory here
 set(sources_list_h
-HyperScore.h
-MorpheusScore.h
-PScore.h
 RNPxlFragmentAnnotationHelper.h
 RNPxlMarkerIonExtractor.h
 RNPxlModificationsGenerator.h
