+++ src/openms/source/ANALYSIS/ID/sources.cmake
@@ -22,6 +22,7 @@ FalseDiscoveryRate.cpp
 FIAMSDataProcessor.cpp
 FIAMSScheduler.cpp
 HiddenMarkovModel.cpp
+HyperScore.cpp
 IDBoostGraph.cpp
 IDConflictResolverAlgorithm.cpp
 IDMapper.cpp
@@ -33,11 +34,13 @@ IDScoreSwitcherAlgorithm.cpp
 IonIdentityMolecularNetworking.cpp
 MessagePasserFactory.cpp
 MetaboliteSpectralMatching.cpp
+MorpheusScore.cpp
 PeptideProteinResolution.cpp
-PrecursorPurity.cpp
-ProtonDistributionModel.cpp
 PeptideIndexing.cpp
 PercolatorFeatureSetHelper.cpp
+PrecursorPurity.cpp
+ProtonDistributionModel.cpp
+PScore.cpp
 SimpleSearchEngineAlgorithm.cpp
 SiriusAdapterAlgorithm.cpp
 SiriusMSConverter.cpp
