+++ src/openms/source/ANALYSIS/RNPXL/sources.cmake
@@ -3,9 +3,6 @@ set(directory source/ANALYSIS/RNPXL)
 
 ### list all filenames of the directory here
 set(sources_list
-HyperScore.cpp
-MorpheusScore.cpp
-PScore.cpp
 RNPxlFragmentAnnotationHelper.cpp
 RNPxlMarkerIonExtractor.cpp
 RNPxlModificationsGenerator.cpp
