+++ src/tests/class_tests/openms/source/MorpheusScore_test.cpp
@@ -9,7 +9,7 @@
 #include <OpenMS/CONCEPT/ClassTest.h>
 
 ///////////////////////////
-#include <OpenMS/ANALYSIS/RNPXL/MorpheusScore.h>
+#include <OpenMS/ANALYSIS/ID/MorpheusScore.h>
 ///////////////////////////
 
 #include <OpenMS/KERNEL/MSSpectrum.h>
