+++ src/tests/class_tests/openms/source/PScore_test.cpp
@@ -11,7 +11,7 @@
 #include <OpenMS/FORMAT/DTAFile.h>
 
 ///////////////////////////
-#include <OpenMS/ANALYSIS/RNPXL/PScore.h>
+#include <OpenMS/ANALYSIS/ID/PScore.h>
 ///////////////////////////
 #include <OpenMS/KERNEL/MSSpectrum.h>
 #include <OpenMS/KERNEL/MSExperiment.h>
